Extract milestone cell builder into shared milestoneCellUtils
Pulls `buildMilestoneCellPayload` and the chapter/book-name helpers out
of `migrationUtils.ts` so importers, migrations, and in-editor
structural edits can share a single source of truth for milestone-cell
shape (label format, edit-history, kind/languageId envelope, optional
initial `metadata.data`).

Pure refactor — no behavior change.
